+Reward commands (/reward, /randomreward, /rewards) +Mode 4 to get value based on weights table. +Method to get random reward by category. +Vault integration and finish coin reward. +Class exp reward. %Fix exp reward. -ClashPoints stuff and shop (Not adding that (yet))

+++ src/main/java/com/clashwars/cwreward/reward/ItemReward.java
@@ -31,7 +31,7 @@ public class ItemReward extends Reward {
         this.item = item;
     }
 
-    public ItemReward(String[] categories, float percentage, CWItem item, Map<Integer, Float> amounts, boolean addTogether) {
+    public ItemReward(String[] categories, float percentage, CWItem item, Map<Integer, Double> amounts, boolean addTogether) {
         super(categories, percentage, amounts, addTogether);
         this.item = item;
     }
